+++ bld/wgml/c/gtxtpool.c
@@ -51,7 +51,7 @@
 /*  allocate / reuse and init a text_chars instance                        */
 /*      optionally fill in text                                            */
 /***************************************************************************/
-text_chars * alloc_text_chars( const char * text, size_t cnt, font_number font )
+text_chars *alloc_text_chars( const char *text, unsigned cnt, font_number font )
 {
     text_chars   *   curr;
     text_chars   *   prev;
